DTSBPS-432: upload-blob

+++ src/integrationTest/java/uk/gov/hmcts/reform/sendletter/util/TestStorageHelper.java
@@ -4,6 +4,7 @@ import com.azure.storage.blob.BlobContainerClient;
 import com.azure.storage.blob.BlobServiceClient;
 import com.azure.storage.blob.BlobServiceClientBuilder;
 import org.testcontainers.containers.DockerComposeContainer;
+import uk.gov.hmcts.reform.sendletter.blob.BlobContainerClientProvider;
 
 import java.io.File;
 
@@ -73,6 +74,17 @@ public class TestStorageHelper {
 
     public void deleteBulkprintContainer() {
         testContainer.delete();
-        encryptedContainer.delete();
+        if (encryptedContainer != null) {
+            encryptedContainer.delete();
+        }
+    }
+
+    public BlobContainerClientProvider getBlobServiceClientProvider() {
+        return containerName -> {
+            var blobServiceClient = new BlobServiceClientBuilder()
+                .connectionString(String.format(STORAGE_CONN_STRING, dockerHost, 10000))
+                .buildClient();
+            return blobServiceClient.getBlobContainerClient(containerName);
+        };
     }
 }
